Refactor code: delete themeMap and use extension methods as replacement
In the previous code themeMap was used to map the string of theme names to the app theme enum. Now two different extensions called toName() and fromName() are used to replace this map and retrieve string of theme from AppName and vise versa. The rest of the code has been refactored to work with this adjustment